// packages/cli-server/generated/cli.schemas.ts
export type CliAuthMode = typeof CliAuthMode[keyof typeof CliAuthMode];


export const CliAuthMode = {
  browser_session: 'browser_session',
  bearer_token: 'bearer_token',
} as const;

export interface CliAuthSessionProjectedUser {
  id?: string;
  /** @pattern ^(?!\.)(?!.*\.\.)([A-Za-z0-9_'+\-\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$ */
  email?: string;
  displayName?: string;
}

export interface CliAuthSessionResponse {
  authMode?: CliAuthMode;
  user?: CliAuthSessionProjectedUser;
  /**
   * @minLength 1
   * @nullable
   */
  activeOrgSlug?: string | null;
  /**
   * @nullable
   * @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
   */
  issuedAt?: string | null;
  /**
   * @nullable
   * @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
   */
  expiresAt?: string | null;
}